requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
chromadb
google-generativeai
sentence-transformers
pysqlite3-binary